More tweakage.
This seems a reasonable middle ground for both PacBio and Illumina.

We're now using a weighted mean of the average qual and maximum qual,
and then using this in the various seqQ/indelQ functions.  This means
things like PB with uncalibrated very high quals and the lower qual
Illumina still play nicely without having to have machine specific
parameters to change all the quality thresholds.  It's not perfect,
but is simple and doesn't introduce more tuning complexity for the
user.